+++ plugins/org.eclipse.sirius.diagram/src-gen/org/eclipse/sirius/diagram/description/style/impl/WorkspaceImageDescriptionImpl.java
@@ -51,7 +51,7 @@ public class WorkspaceImageDescriptionImpl extends NodeStyleDescriptionImpl impl
      * @generated
      * @ordered
      */
-    protected static final Integer ARC_WIDTH_EDEFAULT = new Integer(1);
+    protected static final Integer ARC_WIDTH_EDEFAULT = new Integer(10);
 
     /**
      * The cached value of the '{@link #getArcWidth() <em>Arc Width</em>}'
@@ -71,7 +71,7 @@ public class WorkspaceImageDescriptionImpl extends NodeStyleDescriptionImpl impl
      * @generated
      * @ordered
      */
-    protected static final Integer ARC_HEIGHT_EDEFAULT = new Integer(1);
+    protected static final Integer ARC_HEIGHT_EDEFAULT = new Integer(10);
 
     /**
      * The cached value of the '{@link #getArcHeight() <em>Arc Height</em>}'
